reorganize inference scripts with shared funcs
